refactor(db): extract database foundation (OPE-166)
