+++ packages/db/src/session-tool-call-settlement.ts
@@ -1,7 +1,7 @@
 import type { SessionEvent } from "@opengeni/contracts";
 import { boundModelToolOutputItem } from "@opengeni/codex";
 import { and, asc, eq, sql } from "drizzle-orm";
-import type { Database } from "./index";
+import type { Database } from "./database";
 import { sanitizeEventPayload, sanitizeModelPayload } from "./event-payload-sanitizer";
 import * as schema from "./schema";
